+++ src/useMedia/useMedia.spec.js
@@ -1,8 +1,6 @@
 import { renderHook, act } from '@testing-library/react-hooks'
 import useMedia from './useMedia'
 
-function MatchMediaStub() {}
-
 describe('useMedia', () => {
     describe('fallback', () => {
         const query = '(something: impossible)'