+++ src/utils/isBrowser.node.spec.js
@@ -1,0 +1,11 @@
+/**
+ * @jest-environment node
+ */
+
+import isBrowser from './isBrowser'
+
+describe('isBrowser (node)', () => {
+    it('returns true', () => {
+        expect(isBrowser()).toBe(false)
+    })
+})