+++ src/utils/isBrowser.spec.js
@@ -1,0 +1,10 @@
+/**
+ * @jest-environment jsdom
+ */
+import isBrowser from './isBrowser'
+
+describe('isBrowser', () => {
+    it('returns true', () => {
+        expect(isBrowser()).toBe(true)
+    })
+})